+++ CHANGELOG.md
@@ -13,6 +13,14 @@
 - [**BC**] Move `Option` functions in `Input` module into `Input.Option` module and rename them to be shorter
 - [**BC**] Move `Argument` functions in `Input` module into `Input.Argument` module and rename them to be shorter
 - Add `Help.lines` function to format lines for a command help
+- [**BC**] Replace `ConsoleApplicationError.ConsoleApplicationError` with `ConsoleApplicationError.ConsoleApplicationException`
+- Show full exception stack trace with verbose output
+- [**BC**] Add Execute cases, which must be explicitly declared for an execute function
+    - `Execute.Execute`
+    - `Execute.ExecuteResult`
+    - `Execute.ExecuteAsync`
+    - `Execute.ExecuteAsyncResult`
+- Add `runAsyncResult` function
 
 ## 3.1.0 - 2022-10-17
 - Show multiple errors instead of a first one, if there are more than one